refactor: 패키지명 엔티티를 도메인으로 변경
