+++ src/main/java/roomescape/controller/reservation/ReservationRequest.java
@@ -1,7 +1,7 @@
 package roomescape.controller.reservation;
 
-import roomescape.entity.Reservation;
-import roomescape.entity.ReservationTime;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 
 import java.time.LocalDate;
