+++ src/main/java/roomescape/controller/time/TimeController.java
@@ -2,9 +2,7 @@ package roomescape.controller.time;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -38,14 +36,5 @@ public class TimeController {
                 .status(HttpStatus.OK)
                 .location(URI.create("/times/" + time.id()))
                 .body(time);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTime(@PathVariable("id") final Long id) {
-        int deletedCount = timeService.deleteTime(id);
-        if (deletedCount == 0) {
-            return ResponseEntity.notFound().build();
-        }
-        return ResponseEntity.ok().build();
     }
 }
