+++ src/main/java/roomescape/repository/ReservationTimeH2Repository.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.entity.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 import javax.sql.DataSource;
 import java.sql.ResultSet;
