+++ src/test/java/roomescape/MissionStepTest.java
@@ -11,7 +11,7 @@ import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import roomescape.controller.reservation.ReservationController;
-import roomescape.entity.Reservation;
+import roomescape.domain.Reservation;
 
 import java.lang.reflect.Field;
 import java.sql.Connection;
