+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
