+++ src/test/java/roomescape/repository/ReservationTimeRepositoryTest.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.entity.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 import javax.sql.DataSource;
 import java.time.LocalTime;
